Moved AccessHelper to an outer class
